+++ src/modules/esm.mjs
@@ -1,0 +1,47 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+import { fileURLToPath } from 'url';
+import { readFileSync } from 'fs';
+import { resolve } from 'path';
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = JSON.parse(
+    readFileSync(resolve('src', 'modules', 'files', 'a.json'), 'utf-8')
+  );
+} else {
+  unknownObject = JSON.parse(
+    readFileSync(resolve('src', 'modules', 'files', 'b.json'), 'utf-8')
+  );
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const currentFilePath = fileURLToPath(import.meta.url);
+const currentDirPath = path.dirname(currentFilePath);
+
+console.log(`Path to current file is ${currentFilePath}`);
+console.log(`Path to current directory is ${currentDirPath}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
